Everything is loading fine now
Former-commit-id: 81da046e070ba077e91b0870ed17929bc418f7b4

+++ de.fu_berlin.imp.seqan.usability_analyzer.core/src/de/fu_berlin/imp/seqan/usability_analyzer/core/model/dataresource/FileBaseDataContainer.java
@@ -81,7 +81,7 @@ public class FileBaseDataContainer extends FileDataContainer implements
 	@Override
 	public void resetStaticFile(String scope, String name) throws IOException {
 		File staticFile = getStaticFile(scope, name);
-		if (staticFile.exists())
+		if (staticFile != null && staticFile.exists())
 			staticFile.delete();
 	}
